batch of refactoring Subscription(Controller)Service
